% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompleteGenotypes.R
\name{CompleteGenotypes}
\alias{CompleteGenotypes}
\title{A function to remove missing data}
\usage{
CompleteGenotypes(file, nloci, StartYear = 2008, ...)
}
\arguments{
\item{file}{.csv Input file name with genotypic information}

\item{nloci}{Number of loci in the data file}

\item{StartYear}{Year that the simulation starts}
}
\description{
This function allows you to remove missing data from your file, by sampling from the allele frequency distributions at each locus
}
\examples{
CompleteGenotypes()
}
\keyword{data}
\keyword{missing}
