% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rn.Colony.R
\name{Resume.Rn.Colony}
\alias{Resume.Rn.Colony}
\title{Resume running Colony2 through R for parentage on muliple PseudoBabies simulations}
\usage{
Resume.Rn.Colony(ColonyDir, nSim, Miss_data = FALSE, MissingVec = c(0,
  2.5), Geno_Error = FALSE, ErrorVals, Markerfile,
  ShowProgress = FALSE)
}
\arguments{
\item{ColonyDir}{Directory where the Colony program can be called from}

\item{Miss_data}{Would you like to add missing data to the simulations? Defaults to FALSE}

\item{MissingVec}{A vector of values for the percent of missing data in the input file. Required if Miss_data=TRUE. Defaults to 0 and 2.5 percent.}

\item{Geno_Error}{Would you like to add genotyping error to the simualated data? Defaults to FALSE}

\item{Markerfile}{.csv for marker panels that will be evaluated.}

\item{nSims}{Number of replicate simulations}

\item{ErrorVec}{A vector of error values for genotyping error. Required if Geno_Error=TRUE. Defaults to 0 and 2 percent.}
}
\description{
This function facilitates the analysis of PseudoBabies simulations with the program Colony
}
\examples{
Resume.Rn.Colony()
}
\keyword{Colony}
\keyword{parentage}
