% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rn.FRANzHatch.R
\name{Resume.Rn.FRANzHatch}
\alias{Resume.Rn.FRANzHatch}
\title{Resume running FRANz through R for parentage on muliple PseudoBabies simulations}
\usage{
Resume.Rn.FRANzHatch(
  nSim,
  Miss_data = T,
  MissingVec,
  Geno_Error = T,
  ErrorVals,
  Markerfile = "MarkerPanels.csv",
  ShowProgress = FALSE
)
}
\arguments{
\item{Miss_data}{Would you like to add missing data to the simulations? Defaults to FALSE}

\item{MissingVec}{A vector of values for the percent of missing data in the input file. Required if Miss_data=TRUE. Defaults to 0 and 2.5 percent.}

\item{Geno_Error}{Would you like to add genotyping error to the simualated data? Defaults to FALSE}

\item{Markerfile}{.csv for marker panels that will be evaluated.}

\item{ShowProgress}{Would you like a progress bar to track the analysis. Defaults to FALSE}

\item{nSims}{Number of replicate simulations}

\item{ErrorVec}{A vector of error values for genotyping error. Required if Geno_Error=TRUE. Defaults to 0 and 2 percent.}
}
\description{
This function facilitates the analysis of PseudoBabies simulations with the program FRANz
}
\examples{
Resume.Rn.FRANz()
}
\keyword{FRANz}
\keyword{parentage}
\keyword{resume}
