% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim.MG.Data.R
\name{Sim.MG.Data}
\alias{Sim.MG.Data}
\title{Simulate genotypes over many generations}
\usage{
Sim.MG.Data(AnalysisName, Founders, AOMfile,
  Markerfile = "MarkerPanels.csv", NumuSats, NumSNPs, NB = 600,
  NumYears = 19, StartYear = format(Sys.Date(), "\%Y"), nSim = 10,
  Geno_Error = FALSE, Miss_data = FALSE, MissingVec = c(0, 2.5),
  ErrorVals = 1, Programs = c("FRANz", "Colony"), SexInf = FALSE,
  perl.dir, GPHomogeneity = FALSE, OffspringDist = "Poisson",
  SexInfo = FALSE, MatingStr = "Polygamy", lambda = 2, Umin, Umax,
  nBprob, nBsize, ztnbsize, ztnbprob, ...)
}
\arguments{
\item{AnalysisName}{name of the analysis}

\item{Founders}{.csv file containing the genotypes to use for initiating the simulation. Use function CompleteGenotypes to remove missing data before performing simulations.}

\item{AOMfile}{.csv of Age of maturity estimates.}

\item{Markerfile}{.csv for marker panels that will be evaluated.}

\item{NumuSats}{number of microsatellite markers.}

\item{NumSNPs}{number of SNP markers.}

\item{NB}{number of crosses to perform each year. Defaults to 100.}

\item{NumYears}{How many years should the simulation last? Defaults to 50.}

\item{StartYear}{What year should the simulation start? Defaults to current year.}

\item{Geno_Error}{Would you like to add genotyping error to the simualated data? Defaults to FALSE}

\item{Miss_data}{Would you like to add missing data to the simulations? Defaults to FALSE}

\item{MissingVec}{A vector of values for the percent of missing data in the input file. Required if Miss_data=TRUE. Defaults to 0 and 2.5 percent.}

\item{Programs}{Which programs would you like to create input files for? Defaults to c('FRANz','Colony')}

\item{SexInf}{Use the sex of individuals in the parentage inference program}

\item{perl.dir}{File path for the perl directory for FRANz utility for formating.}

\item{OffspringDist}{What distribution should we use for the number of offspring produced for each cross. Options are Uniform (specify a min and a max), NegativeBinomial (specify a prob for the scale parameter see \code{\link[stats]{NegBinomial}} options), and Poisson (specify lambda).}

\item{SexInfo}{Does your dataset have information on sex?}

\item{MatingStr}{Should the simulation allow for polygamy, polygyny, polandry, or monogomy?}

\item{nSim=10}{How many replicate simulations would you like to do? Defaults to 10.}

\item{ErrorVec}{A vector of error values for genotyping error. Required if Geno_Error=TRUE. Defaults to 0 and 2 percent.}
}
\description{
This function allows you to simulate multilocus genotypes through time
}
\examples{
Sim.MG.Data()
}
\keyword{genetics}
\keyword{parentage}
\keyword{simulation}
