% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim.SG.Data.R
\name{Sim.SG.Data}
\alias{Sim.SG.Data}
\title{Simulate genotypes over a single year}
\usage{
Sim.SG.Data(Founders = "FoundersCompleteGenotypes.csv",
  Markerfile = "MarkerPanels.csv", NumSNPs = 93, NumuSats = 12,
  NB = 650, StartYear = as.numeric(format(Sys.Date(), "\%Y")),
  nSim = 10, Geno_Error = F, ErrorVals, Miss_data = F, MissingVec,
  Programs = c("FRANz", "Colony"), SexInf, perl.dir,
  GPHomogeneity = FALSE, OffspringDist = "Poisson", SexInfo = F,
  lambda, Umin = 0, Umax = 1, nBprob, nBsize, ztnbprob, ztnbsize,
  MatingStr = "Polygyny")
}
\arguments{
\item{Founders}{.csv file containing the genotypes to use for initiating the simulation. Use function CompleteGenotypes to remove missing data before performing simulations.}

\item{Markerfile}{.csv for marker panels that will be evaluated.}

\item{NumSNPs}{number of SNP markers.}

\item{NumuSats}{number of microsatellite markers.}

\item{NB}{number of crosses to perform. Defaults to 100.}

\item{StartYear}{What year should the simulation start? Defaults to current year.}

\item{nSim}{How many replicate simulations would you like to do? Defaults to 10.}

\item{Geno_Error}{Would you like to add genotyping error to the simualated data? Defaults to FALSE}

\item{ErrorVals}{How many different values of genotyping error for simulations. Required if Geno_Error=TRUE. Values specified in Loci_error.csv}

\item{Miss_data}{Would you like to add missing data to the simulations? Defaults to FALSE}

\item{MissingVec}{A vector of values for the percent of missing data in the input file. Required if Miss_data=TRUE.}

\item{Programs}{Which programs would you like to create input files for? Defaults to c('FRANz','Colony')}

\item{SexInf}{Use the sex of individuals in the parentage inference program}

\item{perl.dir}{Full path to .csv.pl perl script distributed with FRANz for formatting input files. Required when FRANz is specified as a Program argument}

\item{GPHomogeneity}{A logical to indicate if genepop should be run to compare the allele frequency distribution for all simulated datasets with the founders dataset. Defaults to FALSE}

\item{OffspringDist}{What distribution should we use for the number of offspring produced for each cross. Options are Uniform (specify a min and a max), NegativeBinomial (specify a prob for the scale parameter see \code{\link[stats]{NegBinomial}} options), and Poisson (specify lambda).}

\item{SexInfo}{Does your dataset have information on sex?}

\item{MatingStr}{Should the simulation allow for polygamy, polygyny, polandry, or monogamy?}
}
\description{
This function allows you to simulate multilocus genotypes under the assumption of completely random mating for a single year
}
\examples{
Sim_SG_Data()
}
\keyword{genetics}
\keyword{parentage}
\keyword{simulation}
