% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SumFRANz.R
\name{SumFRANz}
\alias{SumFRANz}
\title{Summarize the results from FRANz analysis for parentage on muliple PseudoBabies simulations}
\usage{
SumFRANz(nSim, Miss_data = FALSE, MissingVec = c(0, 2.5),
  Geno_Error = FALSE, ErrorVals, Cutoff, Markerfile)
}
\arguments{
\item{Miss_data}{Would you like to add missing data to the simulations? Defaults to FALSE}

\item{MissingVec}{A vector of values for the percent of missing data in the input file. Required if Miss_data=TRUE. Defaults to 0 and 2.5 percent.}

\item{Geno_Error}{Would you like to add genotyping error to the simualated data? Defaults to FALSE}

\item{Cutoff}{cutoff value for probability value accepting parentage.}

\item{Markerfile}{.csv for marker panels that will be evaluated.}

\item{nSims}{Number of replicate simulations}

\item{ErrorVec}{A vector of error values for genotyping error. Required if Geno_Error=TRUE. Defaults to 0 and 2 percent.}
}
\description{
This function summarizes the results from multiple runs of Colony2 for parentage inference from PseudoBabies simulations.
}
\examples{
SumFRANz()
}
\keyword{FRANz}
\keyword{parentage}
