% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summarize.LOD.R
\name{Summarize.LOD}
\alias{Summarize.LOD}
\title{Summarize the LOD scores from parentage inference with FRANz}
\usage{
Summarize.LOD(AnalysisName, nSim, Miss_data = FALSE, MissingVec = c(0,
  2.5), Geno_Error = FALSE, ErrorVec = c(0, 2), SavePlot = FALSE,
  PlotName = "LODsum", Plots = c("LCA", "PLCA", "PLCA", "PLICA"),
  ErrorLab, Markerfile, ...)
}
\arguments{
\item{AnalysisName}{Name of the Analysis}

\item{Miss_data}{Would you like to add missing data to the simulations? Defaults to FALSE}

\item{MissingVec}{A vector of values for the percent of missing data in the input file. Required if Miss_data=TRUE. Defaults to 0 and 2.5 percent.}

\item{Geno_Error}{Would you like to add genotyping error to the simualated data? Defaults to FALSE}

\item{ErrorVec}{A vector of error values for genotyping error. Required if Geno_Error=TRUE. Defaults to 0 and 2 percent.}

\item{SavePlot}{Do you want to create plots? Defaults to FALSE}

\item{PlotName}{What do you want to name your plots? Defaults to LODsum}

\item{Plots}{What plots do you want to create? Options are 'LCA','PLCA','PLCA','PLICA'.}

\item{ErrorLab}{Label for Error in plots}

\item{Markerfile}{.csv for marker panels that will be evaluated.}

\item{nSims}{Number of replicate simulations}
}
\description{
This function summarizes the LOD scores from parentage inference of PseudoBabies simulations with the software FRANz.
}
\examples{
summarize.LOD()
}
\keyword{FRANz}
\keyword{LOD}
\keyword{parentage}
